refactor: move rpc to its own module
The `node_control`, `faucet_control` and `daemon_control` modules will shortly be removed since we
don't need duplicate service management code. With that in mind, we move the RPC code from
`daemon_control` to its own module.
