+++ sn_service_management/src/daemon.rs
@@ -85,7 +85,7 @@ impl ServiceStateActions for DaemonService {
     }
 
     fn pid(&self) -> Option<u32> {
-        self.service_data.pid.clone()
+        self.service_data.pid
     }
 
     fn on_remove(&mut self) {
