+++ sn_service_management/src/faucet.rs
@@ -92,7 +92,7 @@ impl ServiceStateActions for FaucetService {
     }
 
     fn pid(&self) -> Option<u32> {
-        self.service_data.pid.clone()
+        self.service_data.pid
     }
 
     fn on_remove(&mut self) {
